+++ src/main/java/com/marginallyclever/robotoverlord/components/GCodePathComponent.java
@@ -4,11 +4,10 @@ import com.jogamp.opengl.GL2;
 import com.marginallyclever.convenience.helpers.OpenGLHelper;
 import com.marginallyclever.convenience.PrimitiveSolids;
 import com.marginallyclever.robotoverlord.SerializationContext;
-import com.marginallyclever.robotoverlord.entity.Entity;
 import com.marginallyclever.robotoverlord.parameters.DoubleParameter;
 import com.marginallyclever.robotoverlord.parameters.IntParameter;
 import com.marginallyclever.robotoverlord.parameters.StringParameter;
-import com.marginallyclever.robotoverlord.swinginterface.componentmanagerpanel.ViewElementSlider;
+import com.marginallyclever.robotoverlord.parameters.swing.ViewElementSlider;
 import com.marginallyclever.robotoverlord.systems.render.gcodepath.*;
 import org.json.JSONException;
 import org.json.JSONObject;
