+++ src/main/java/com/marginallyclever/robotoverlord/systems/robot/cartesian/MakeCartesian3.java
@@ -10,7 +10,7 @@ import com.marginallyclever.robotoverlord.components.OriginAdjustComponent;
 import com.marginallyclever.robotoverlord.components.shapes.MeshFromFile;
 import com.marginallyclever.robotoverlord.systems.render.mesh.load.MeshFactory;
 import com.marginallyclever.robotoverlord.parameters.StringParameter;
-import com.marginallyclever.robotoverlord.swinginterface.componentmanagerpanel.ViewElementFilename;
+import com.marginallyclever.robotoverlord.parameters.swing.ViewElementFilename;
 import com.marginallyclever.robotoverlord.systems.OriginAdjustSystem;
 
 import javax.swing.*;
